Improve UrlBar/UrlSelect functionality (#1143)
This PR improves the UrlSelect component responsible for navigating
through the app's path history. Previously, the feature functioned like
a `<select>`, only allowing traversal to locations that have been
visited before.

Now, the user will be able to enter any path, valid or not, which will
be passed to the emulator and handled by the application. The old
dropdown has been repurposed as a suggestion and history panel, similar
to what we see in modern web browsers. In the future, the panel is
planned to list all available static paths, too.

Recording of the functionality:
https://github.com/user-attachments/assets/21f01e74-50e0-4fc5-9e38-512dfe4419bd

### How Has This Been Tested:
- Open any multi-screen RN project using Expo Router or React Navigation
- The path bar becomes enabled once the app is fully loaded
- Any changes in navigation invoked from the emulator will be instantly
visible in the bar
- Click on the path field - it should become active and open the
suggestions
- Enter the desired path and press Enter/Return - the app should
navigate to the location or display a "404" page if it's invalid
- As you're writing, suggestions from the history should appear below
- When the field is active, press the down arrow on your keyboard - the
first option from the suggestion panel should become focused - confirm
with Enter/Return, navigate with up/down arrows
- When the first option from the list is highlighted, pressing the up
arrow should focus the input field
- Open a component preview from the editor window - the app should
display it and a path starting with `preview://` should show up on the
path bar
- The preview URL appears in the history and is navigable like any other
URL