Make lint properly run and fix accumulated errors (#1152)
After updating dependencies in #1107 we changed eslint config to adapt
it to a new version in a way that made lint skip all of the files we had
in the codebase and effectively we stopped running lint altogether.

This PR restores the lint to run. Since the defaults changed, we are now
changing the behaviors of some rules that are specifically annoying but
also updating code to reflect rules that are important.

### How Has This Been Tested: 
1. Run lint, see it passes
2. Test critical changes, specifically in `waitForEmulatorOnline`