improve build error alert messages (#1157)
After #1068, the build error alerts display the error messages of the
exceptions thrown while starting the app.
This resulted in the IDE showing non-user-friendly messages to the user.
This PR wraps some of these errors with user-friendly messages
redirecting the user to the logs for details.

### How Has This Been Tested: 
- break the native code (`ios`/`android` directory) in order to cause a
build error
- verify the message redirects you to the build logs to check for
details
- break your node install and restart the IDE
- verify the error message when installing node modules redirects you to
the Radon logs