Simplify settings management by moving it out of device session manager (#1159)
This PR extracts fragments of the changes made in #1154 that simplifies
the management of device settings.

Before, we'd store settings in device session which is already a pretty
complex class with a dozen of different responsibilities. Eventually,
the device settings need to be passed to individual device, so I figured
it'd be simpler to move them there avoiding extra logic to be held in
the session.

This PR refactors the code such that we move the responsibility of
managing the device settings onto the base device class that android and
iOS device subclesses inherit. This allows the subclasses to access
settings when needed as protected class member. The base class also
handles the settings that apply only to the preview which is managed by
the base device class already (like replays or display touches).

As a consequence, we also refactor project class a bit. We don't need
device change callback, as the change can only happen as a result of the
project call.

TODO: For supporting parallel sessions, we'd need to implement
additional mechanism for monitoring device session updates that happen
to the active device and apply them onto the inactive one. This is out
of the scope of the current change as it'd require modification to the
way we activate sessions and is also something that we doesn't handle
right now.

### How Has This Been Tested: 
1. Boot device, change settings (show touches in particular), switch to
a different device and make sure the changes are reflected.
2. Reboot the IDE to see the changes are persistent