Add approot select to Preview (#1162)
Adds a dropdown to the main Preview screen which allows the user to
quickly switch between applications in a monorepo as well as see at a
glance which app root is selected.

### Known issues:
- sometimes after switching the AppRoot, a "Couldn't start selected
device" error appears. This doesn't prevent the app to build and run
correctly (and after it does, Radon will stop showing the error and work
normally), so it seems to be an issue with error handling in device
startup.

### TODO:
- [x] ~cleanup css classes in webview code~ I gave up
- [x] improve how application's name is derived (support
`app.config.js`, use `package.json` as a fallback)

![Screenshot 2025-05-20 at 14 26
06](https://github.com/user-attachments/assets/7898979d-54ca-4ebd-a7d1-a6cb06a531c2)
![Screenshot 2025-05-20 at 14 27
26](https://github.com/user-attachments/assets/254717f0-04dc-41c2-9c0a-591c80c60e3c)


### How this was tested:
- open a monorepo or a folder containing multiple RN apps
- the half-pill button next to the device picker should display the name
of the current application
- press the button
- the dropdown should contain all the RN apps in the workspace
- selecting an entry should switch Radon to that app