Refactor project and device state to fix issues with device switch and simplify our logic (#1154)
This PR refactors a bunch of places in order to straighten up the logic
and data flow and to make the project state management more robust and
self-contained.

Key objectives:
1) Switching between devices sometimes causes an undesirable state
coming from unselected/old device to leak into project settings. As this
is undesirable, we update `DeviceSessionManager` to guard against this.
Now, all the project state updates are routed via `DeviceSessionManager`
which verifies that the device session from which the update comes is
the one that is currently selected.
2) Our state is currently distributed between the extension process and
webview. Most of the "project state" actually belongs to a particular
device session. When we switch between sessions, we don't want to
manually keep track of the state updates on the webview site. This was
previously happening. When switching we'd need to reset navigation
state. But this also applies to other state elements, like whether we're
currently profiling or if there's a stale build. Previously, we'd use
events to pass that information and then store state in webview. Now,
we're moving most of that into the "project state" and update it along
all the other variables. This includes things like navigation history
which is now kept in the state too.

The biggest change in the PR that affects the biggest number of lines is
that we're moving most of the "project state" out of the project class
and into device session class. This includes a number of methods and
logic responsible for monitoring state updates and responding to them.

This PR also aims to fix some issues and design flaws in project /
device-session-manager / device session classes. Below, we're listing
notable changes:

1. device settings management moved from device session into device
base. Now, device base class is responsible for storing the updates for
device settings and also managing replays and show touches settings.
2. we're restructuring the startup flow for the device session class.
Before it'd control all the aspects of starting the session. Since we
recently added "activate" / "deactivate" methods that aim to control the
session "visibility" when switching, we move control over the debug
session there. We also branch out from the startup flow to make sure
that debugger is only started when in active state.
3. we're removing isolated "tools state". the "tools state" is now under
the device session state and is surfaced via "project state" (I know it
is confusing, but trust me, this leads to a better structure)
4. we're restructuring navigation handling. Now the navigation history
is also a part of the device session state. We add extra methods like
"navigateHome" and "navigateBack" such that the device session class can
control the updates.


Future work:
We'd ideally want to have a logical distinction between what's in the
project state and what's in the individual device session state. For the
future, we plan to extract the relevant portions to separate "states".
But in order to avoid more boilerplate we'd like to switch to some
state-management solution that we can control from extension and that
syncs with the webview code (i.e. zustand).


### How Has This Been Tested: 
1. Run app (test on several versions)
6. Switch devices during boot
7. Test restart options
8. Switch during restart
9. After switching verify there are no multiple debug sessions in the
debug console dropdown
10. Test Url bar on expo-router project
11. Verify changing device settings works and persist across device
switching
12. Test replays and "show toucches" settings in particular