Fix and restore ensureDependenciesAndNodeVersion functionality  (#1170)
After #1154 `ensureDependenciesAndNodeVersion` was not passed into
device session, this PR changes that and restores the functionality it
provided.

Additionally investigation into #1163 revealed a problem withhoisted
pnpm monorepos namely dependencies installed directly in `node_modules`
and not in `node_modules/.pnpm` are reported as `unsavedDependencies` by
`pnpm ls` command, which caused us to reinstall pnpm every time,
potentially causing problems with the set up.

The changes made to the `isPnpmModulesInstalled` function change the
assumption that `node_modules` are installed only if `dependencies` are
detected and also allow `unsavedDependencies` which prevents unnecessary
installs in hoisted setups.

### How Has This Been Tested: 

- run `expo-52` template application in pnpm monorepo, with
`node-linker=hoisted` setting in `.npmrc`