Fix Radon not starting when no devices are found (#1173)
This PR fixes an issue with radon not loading when no devices are
available.

The problem was caused by `initialized` in `projectState` not being
updated as after #1154 `initialized` is only updated when
`updateProjectState` and it does not happen if no devices are found.

To solve this problem we introduce a new `deviceSessionsManagerDelegate`
method and update `initialized` even when no device can be selected

### How Has This Been Tested: 

- Run radon with no devices


note: 

this solution essentially returns to the way `initialized` was updated
before #1154 but ideally a selection process should happen on the
webview side, which will allow us to avoid this problem entirely.