Extract route information from Expo Router and display it in UrlSelect (#1160)
This PR further enhances the extension's integration with Expo Router by
automatically extracting all indexed paths and displaying them in the
existing URL suggestion panel. Now, the user doesn't need to navigate to
a path for it to become listed, which significantly improves the
usability of the navigation bar.

Paths visited recently, including previews and invalid URLs, are
displayed above them for easy access. The indexed routes are sorted by
depth, and then alphabetically, ensuring the paths closer to the app
root will always stay on top.

As in big apps the list can get really long, the user can type into the
bar to filter the routes and navigate to them either with mouse or
keyboard. The suggested paths now highlight the query substring to
improve readability.

The component identifies and highlights dynamic route segments, and if a
route with one or more such segments is selected, the user will be able
to fill them in the input field before navigating.

### How Has This Been Tested: 
- open a React Native project with Expo Router
- wait until the app loads and click on the URL field - the list now
contains all URLs available in the app
- type something into the text field - paths will be suggested and
highlighted based on the query
- if the app has dynamic routes, the parts in `[]` are colored, and the
path is editable in the text field upon click - the first dynamic
segment will be auto-selected, and the user can use (Shift+)Tab to jump
between them
- try to make the bar content overflow - now the list becomes scrollable
without breaking the layout



https://github.com/user-attachments/assets/283fee9a-56d3-4f12-9cf3-774a546c50d7