Fix bundle error popup after fast refresh (#1175)
This PR fixes an issue with bundle error popup either flashing or
completely not showing when bundle error is caused be "fast refresh".
The problem was happening because even if fast refresh is unsuccessful
it is reported and coupled which means that `RNIDE_fastRefreshComplete`
event does not mean that the app is running.

To avoid this issue we restore additional checks from before  #1154. 

### How Has This Been Tested: 

- run `react-native-77` test app and cause bundle error with fast
refresh