Fix network panel scroll on start and blinking column borders (#1181)
This PR fixes two visual issues in the network panel:
1) When network panel is just opened, the content isn't scrollable until
you resize the panel pane
2) When hovering over the requests lists, we'd get the column bar
appearing and disappearing

The first issue is fixed by triggering the resize change on component
mount.

The second issue is fixed by adding `bordered-columns` prop to
`VscodeTable` component. Apparently, this option is require for
`resizable` tables, as otherwise the columns only appear when you try to
"resize" them which is when you hover. It look better when the columns
are shown at all times in that case.

### How Has This Been Tested: 
1. Open network panel
2. Make it show a bunch of requests, make sure you can scroll the
payload window w/o previously resizing the editor window or the panel
pane
3. Make sure the column borders are shown at all time on the requests
list.