Fix launch config changes reloading the application (#1197)
Currently, any changes to the `launch.json` file cause the application
to be reloaded when no custom `appRoot` is set in the config.
This is caused by the launch config change listener to always setup a
new app root which causes the `DeviceSessionManager` to be disposed,
even if the app root didn't actually change.
This PR changes that behavior to instead check if the actual app root
changed and only reset in that case.

### How Has This Been Tested: 
- open a project without an `appRoot` specified in `launch.json`
- change the contents of `launch.json`
- the application should not restart