Fix UrlSelect item focusing and unreadable dynamic segment colors (#1193)
This PR fixes two bugs noticed in the behavior of the `UrlSelect`
component:
- the keyboard item focusing functionality sometimes breaking after
adding a new path to the dropdown, e.g. after opening a preview - caused
by duplicated entries in the array of item references
- ~the suggestions extracted from Expo Router not showing after the app
started, especially on Android - caused by incorrect logic of the
DevTools agent listener, which led to the route list not being sent to
the webview~
fixed in #1180

Additionally, I made the dynamic segments in the dropdown more visible
by changing `color: var(--swm-disabled-text)` to a new, custom color.

### How Has This Been Tested: 
1. Open a multi-route Expo Router project with Expo version ≤52 and
launch it on an Android device
2. After the app loads, click on the `UrlSelect` - the indexed routes
should be displayed properly and with dynamic segments brightly colored
3. Navigate to some route, then open a preview, navigate back to `/`,
play with the filtering in the bar while testing the arrow key focusing
at each point - it shouldn't get stuck or focus incorrect items anymore