Add an option to remove navigation history entries (#1219)
This PR improves the functionality of the `UrlSelect` by allowing the
user to remove unwanted entries (e.g. with typos) from the navigation
history, since we don't clear it after reload anymore.

If the removed item is a static route present in the `routeList`
extracted from Expo Router, it will be moved down to the
alphabetically-sorted "other" section, and all other items will just be
removed. The currently selected route can't be removed until another
route is opened, as this would break the history.


https://github.com/user-attachments/assets/35c745cb-83af-48dc-8eb1-172c1bb493fc

### How Has This Been Tested: 
1. Open an Expo Router app
2. Play with navigation, open some static, dynamic and incorrect routes
3. Open the UrlSelect, there should be an "X" next to each visited entry
4. Click the "X"-es on the items - they should disappear or move down in
the suggestions
5. Refresh the app - the deleted items shouldn't show up until visited
again