Allow running devices in the background (#1231)
Allow the user to leave devices running in the background, to enable
switching quickly between them without having to wait for the device and
application to load.

This PR adds UI elements to allow for managing running devices:
- "Select device" dropdown now displays a separate "Running devices"
section when >1 device is running.

![image](https://github.com/user-attachments/assets/bea62c1a-1958-4140-b614-14a3a7bf4408)

- an indicator next to the "Select device" dropdown showing the number
of devices running in the background (if any)
<img width="354" alt="image"
src="https://github.com/user-attachments/assets/da818966-8721-4054-a975-a1d156ad6e9d"
/>

- In the "Manage Devices" modal, a "Shut down devices when switching"
toggle was added to restore the previous behaviour if a user doesn't
need devices to run in the background
- In the "Manage Devices" modal running devices have a "stop" button to
shut down a single device.

![image](https://github.com/user-attachments/assets/2aa7462d-4013-44b8-8920-d80918290ce1)

- In the "Manage Devices" modal you can now press the device entry to
select it (which saves us from adding too many buttons for devices
running in the background, after adding the "stop" button).

### How Has This Been Tested: 
- open an app in Radon
- go to "Manage Devices" and uncheck "Shut down devices when switching"
- select a different device
- verify the device switches correctly and the app starts
- select the previous device
- verify both running devices are shown in the "Running devices" section
in the device selector
- verify the device switches and the app appears immediately, without
having to boot/build/install/etc.
- check that stopping the device from "Manage devices" works correctly
- check that checking "Shut down devices when switching" restores the
previous behaviour