fix failing to build iOS workspaces with multiple apps (#1240)
Allows Radon to build iOS apps with mulitple xcode projects present in
the `ios` directory.

We change the logic for picking the xcode project/workspace to pass to
xcode build to use `react-native-cli` package to ensure we're able to
build and run apps that regular `npm run ios` would handle.

Fixes #1230

### How Has This Been Tested: 
- open a (non-expo) RN workspace in XCode
- add a project to the workspace and put it inside the `ios` directory
- open the same project in Radon
- it should build