Prevent running the same build concurrently (#1276)
After #1118 we encounter an issue where, if two devices (with the same
platform) are launched at the same time, and we need to build the native
application, both will start building at the same time. This causes
issues (at least on iOS, running `xcodebuild` from the same project
concurrently causes errors in one of the builds) as well as is simply
wasteful (since the application build can be used for both devices
anyway).

This PR introduces a "batching" mechanism, where if a device requests an
application build while a compatible build is already running, it will
use the result of that in-progress build instead.

### How Has This Been Tested: 
- open an app (which builds the native part locally) in Radon
- choose the "Clean rebuild" option to force a build
- start a second device with the same platform
- check that the build only runs once -- either by inspecting the Radon
IDE output or adding breakpoints/logpoints in `BuildManager`
- check that stopping all of the devices also cancels the build