Fix new Android emulators getting stuck in "booting" state (#1282)
When trying to call `adb` on a newly created Android emulator, it
sometimes fails with "device not found" error.
This can cause Radon to be stuck in the "Booting device" state until the
device is restarted by the user, due to an `adb` call made in
`waitForEmulatorOnline`.
This PR tries to work around the issue by retrying the command until a
timeout is reached.
It also throws an error if the `waitForEmulatorOnline` call fails,
instead of getting stuck in "Booting Device".

### How Has This Been Tested: 
- create a new Android device and select it immediately
- wait for the device to start
- on main branch, the preview will be stuck at "Booting device" and you
can see
```
[error] Error: Command failed with exit code 255: adb -s emulator-5556 wait-for-device shell while [[ -z $(getprop sys.boot_completed) ]]; do sleep 0.5; done; input keyevent 82
    at makeError (/Users/jwajgelt/Developer/radon/radon-ide/packages/vscode-extension/node_modules/execa/lib/error.js:60:11)
    at handlePromise (/Users/jwajgelt/Developer/radon/radon-ide/packages/vscode-extension/node_modules/execa/index.js:118:26)
    at processTicksAndRejections (node:internal/process/task_queues:105:5) Unhandled promise rejection
```
in the logs