Setup CDP-based communication mechanism for Radon Connect (#1281)
This PR implements the "agent" interface for Radon Connect.

Agent interface is what we use for communication between the extension
and the app process. When running metro inside the extension, we
implement the agent based on the react devtools websocket connection.
This is then used by the dev tools extensions like redux, react query
and others in order to exchange messages.

In Radon Connect, we don't have a way to communicate over the devtools
protocol, because we don't control metro proces and hence we don't also
control the react devtools port.

The new mechanism relies on the CDPs "binding" mechanism. Using this
mechanism, we can register a global method (we call it
`__radon_binding`), which when called, will trigger
`Runtime.bindingCalled` method. We use this mechanism to trigger message
calls from the app process to the extension (via debugger).

To send messages from extension to app process, we simply use `evaluate`
and call another global method that we need to register beforeahead
(`__radon_dispatch`).

Similarly to how we do things in "Radon Classic", we also need to have
some runtime code executed. The runtime code is responsible for setting
up the dispatch command and defining the agent interface that uses
dispatch command and the registered binding. In order to load the
runtime code however, we can no longer inject some extra modules into
metro. Instead, we again use `evaluate` mechanism. But, because we want
to eventually also load a bit more code beyond the agent definition (for
example to include the redux dev tools code), we are using a pre-built
script that we load into the eval method.

The script is built using esbuild and mostly follows the same pipeline
as the main extension code, apart from the fact we need a slightly
different configuration, because the bundled code will be loaded into a
React Native process rather than a node process.

Here's the breakdown of the changes made:
1) We add a new `connect_runtime.ts` file which is an entry point for
the runtime loaded into RN process when Radon Connect is connected
2) We update esbuild configuration to build the runtime bundle
3) We update the debug session code, such that (when requested) it sets
up the Radon Connect runtime and installs necessary bindings
4) We refactor Connector.ts by defining new ConnectSession class that
now manages the debug session along with metro etc. This simplifies the
Connector code a bit while also allowing us to expand it for handling
tools in the future.

This change has mostly no effect as now we don't utilize the
communication channel (this will be needed once we add tools support).
The only effect this change has is on Radon Connect runtime where we
load the runtime bundle upon connecting.

### How Has This Been Tested: 
1) Since this change has no effect apart from loading runtime code, the
way I tested it is by checking the communication channel. For that you
need to add some lines to `connect_runtime.ts`:
2) add `globalThis.__radon_agent.postMessage("Hello");` and place a
breakpoint in `ConnectSessionDelegate.onBindingCalled` to see that the
event gets delivered there
3) add `globalThis.__radon_agent.onMessage` handler and call
`inspectorBidge.send` to see that the mesages are delivered both ways.